# tacs.api/src/main/resources/swagger/api.yaml
openapi: 3.0.1
info:
  title: TP TACS
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8080/api/v1
...
paths:
...
  /admin/estadisticas:
    get:
      tags:
        - Admin
      summary: Obtiene las estadísticas generales del juego
      operationId: getEstadisticas
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: fechaInicio
          schema:
            type: string
            format: date
          required: false
        - in: query
          name: fechaFin
          schema:
            type: string
            format: date
          required: false
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadisticasDeJuegoModel'
        400:
          description: Bad request
        403:
          description: Forbbiden
  /admin/scoreboard:
    get:
      tags:
        - Admin
      summary: Retorna los datos del scoreboard ordenados por partidas ganadas descendentemente.
      operationId: getScoreboard
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: tamanioPagina
          schema:
            type: integer
            format: int64
            default: 10
          required: false
        - in: query
          name: pagina
          schema:
            type: integer
            format: int64
            default: 0
          required: false
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                title: scoreboardResponse
                properties:
                  scoreboard:
                    type: array
                    items:
                      $ref: '#/components/schemas/EstadisticasDeUsuarioModel'
        400:
          description: Bad request
        403:
          description: Forbidden
        404:
          description: Page not found
  /admin/estadisticas/usuarios/{idUsuario}:
    get:
      tags:
        - Admin
      summary: Obtiene las estadísticas particulares de un usuario
      operationId: getEstadisticasDeUsuario
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: idUsuario
          description: Id del usuario
          schema:
            type: integer
            format: int64
          required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadisticasDeUsuarioModel'
        400:
          description: Bad request
        # En el caso de que no se tengan permisos, también se retorna 404
        404:
          description: Page not found
...
components:
  schemas:
...
    UsuarioModel:
      type: object
      description: Un usuario.
      required:
        - id
        - nombreDeUsuario
      properties:
        id:
          type: integer
          format: int64
        nombreDeUsuario:
          type: string
...
    EstadisticasDeJuegoModel:
      type: object
      properties:
        partidasCreadas:
          type: integer
          format: int64
        partidasEnCurso:
          type: integer
          format: int64
        partidasTerminadas:
          type: integer
          format: int64
        partidasCanceladas:
          type: integer
          format: int64
    EstadisticasDeUsuarioModel:
      type: object
      properties:
        usuario:
          $ref: '#/components/schemas/UsuarioModel'
        partidasJugadas:
          type: integer
          format: int64
        partidasGanadas:
          type: integer
          format: int64
        rachaActual:
          type: integer
          format: int64
          description: Racha actual de victorias
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT